[v0.90][WP-14] Split internal command test responsibilities (Closes #2110)
